+++ app/src/main/java/com/nicolas/freegames/data/local/datasource/GameLocalDataSource.kt
@@ -1,0 +1,12 @@
+package com.nicolas.freegames.data.local.datasource
+
+import com.nicolas.freegames.data.local.entities.GameEntity
+
+interface GameLocalDataSource {
+
+    suspend fun create(gameEntity: List<GameEntity>)
+
+    suspend fun getAll(): List<GameEntity>
+
+
+}
